Refactor move ConfigurationInterface
