+++ src/Anneck/Game/Configuration/WorldConfiguration.php
@@ -12,7 +12,16 @@
 namespace Anneck\Game\Configuration;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Anneck\Game\ConfigurationInterface;
 
+/**
+ * The WorldConfiguration
+ *
+ * @package Anneck\Game\Configuration
+ * @todo    Write PHPDoc for this class!
+ * @since   0.0.1-dev
+ * @author  André Anneck <andreanneck73@gmail.com>
+ */
 class WorldConfiguration implements ConfigurationInterface
 {
     /**
